add first version of "pdb2tikz"
it reads in a molecule from a pdb file and converts it to tikz instruction (tex code) in a "ball and sticks" format, which compiles to ps, eps or pdf.
the bonds are not yet pretty, just lines right now.
todo: caclulate ball-cyclinder-intersection and draw corresponding arc... in the correct z-order
note: you need openbabel dev libaries for this to work